# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZU Connect API - منصة طلبة جامعة الزاوية
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  totalStudents: number;
  totalColleges: number;
  totalActivities: number;
  totalLibraryFiles: number;
  totalCourses: number;
  totalMembers: number;
}

export interface NewsItem {
  id: number;
  title: string;
  body: string;
  category: string;
  date: string;
  viewCount: number;
}

export interface NewsInput {
  title: string;
  body: string;
  category: string;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  instructor: string;
  duration: string;
  level: string;
  category: string;
  totalSeats: number;
  enrolledCount: number;
  colorScheme: number;
}

export interface CourseInput {
  title: string;
  description: string;
  instructor: string;
  duration: string;
  level: string;
  category: string;
  totalSeats: number;
  colorScheme: number;
}

export interface EnrollmentInput {
  studentId: string;
}

export interface Member {
  id: number;
  name: string;
  role: string;
  department: string;
  year: string;
  category: string;
  initials: string;
}

export interface College {
  id: number;
  name: string;
  studentCount: number;
  hasNews: boolean;
  hasSchedules: boolean;
  hasFiles: boolean;
  hasActivities: boolean;
  icon: string;
}

export interface LibraryResource {
  id: number;
  title: string;
  subtitle: string;
  type: string;
  rating: number;
  downloadCount: number;
  college: string;
}

export interface PlannerEvent {
  id: number;
  title: string;
  description: string;
  date: string;
  month: string;
  icon: string;
}

export interface ChatRoom {
  id: number;
  name: string;
  description: string;
  onlineCount: number;
  icon: string;
}

export interface ChatMessage {
  id: number;
  roomId: number;
  sender: string;
  message: string;
  isMe: boolean;
  createdAt: string;
}

export interface ChatMessageInput {
  sender: string;
  message: string;
  isMe: boolean;
}

export interface SuggestionInput {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  college?: string | null;
  type: string;
  message: string;
}

export interface Suggestion {
  id: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  college?: string | null;
  type: string;
  message: string;
  createdAt: string;
}

export interface VolunteerInput {
  name: string;
  college: string;
  phone: string;
  area: string;
}

export interface Volunteer {
  id: number;
  name: string;
  college: string;
  phone: string;
  area: string;
  createdAt: string;
}

export interface FaqItem {
  id: number;
  question: string;
  answer: string;
  category: string;
}

export type ListNewsParams = {
category?: string;
};

export type ListCoursesParams = {
category?: string;
};

export type ListMembersParams = {
category?: string;
};

export type ListLibraryParams = {
type?: string;
};

export type ListPlannerParams = {
month?: string;
};