# lib/api-zod/src/generated/types/college.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZU Connect API - منصة طلبة جامعة الزاوية
 * OpenAPI spec version: 0.1.0
 */

export interface College {
  id: number;
  name: string;
  studentCount: number;
  hasNews: boolean;
  hasSchedules: boolean;
  hasFiles: boolean;
  hasActivities: boolean;
  icon: string;
}