# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZU Connect API - منصة طلبة جامعة الزاوية
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}